rewired RootContainer and renamed files to .jsx
