+++ app/components/Footer.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Footer = () => (
+  <div className="col-lg-12">
+    <h1>This is footer component</h1>
+  </div>
+)
+
+export default Footer
