+++ app/containers/CategoriesContainer.jsx
