+++ app/containers/RootContainer.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Navbar from '../components/Navbar'
 import Sidebar from '../components/Sidebar'
 import FeaturedProducts from '../components/FeaturedProducts'
-
+import Footer from '../components/Footer'
 import CategoriesContainer from './CategoriesContainer'
 
 export default class RootContainer extends Component {
@@ -19,6 +19,8 @@ export default class RootContainer extends Component {
         <div className="row">
           <Sidebar />
           <CategoriesContainer />
+          { this.props.children }
+          <Footer />
         </div>
       </div>
     )
